project: move files around to prepare for the frontend
