+++ tools/place-compress/src/main.rs
@@ -1,7 +1,7 @@
-use gzp::{deflate::Gzip, ZBuilder, Compression};
+use anyhow::Result;
+use gzp::{deflate::Gzip, Compression, ZBuilder};
 use rayon::prelude::*;
 use std::io::prelude::*;
-use anyhow::Result;
 
 fn main() -> Result<()> {
     let mut args = Vec::new();
@@ -15,18 +15,29 @@ fn main() -> Result<()> {
     paths.sort();
 
     eprintln!("starting processing...");
-    let mut encoder = ZBuilder::<Gzip, _>::new().compression_level(Compression::best()).num_threads(0).from_writer(std::fs::File::create("data.bin")?);
+    let mut encoder = ZBuilder::<Gzip, _>::new()
+        .compression_level(Compression::best())
+        .num_threads(0)
+        .from_writer(std::fs::File::create("data.bin")?);
     for chunk in paths.par_iter().chunks(500).collect::<Vec<_>>() {
         eprintln!("processing...");
-        let data: Vec<u8> = chunk.par_iter()
+        let data: Vec<u8> = chunk
+            .par_iter()
             .map(|path| {
                 eprintln!("{}", path.display());
                 image::open(path)
             })
             .filter_map(Result::ok)
             .map(|image| {
-                image.to_rgba8().enumerate_pixels().map(|(_, _, pixel)| pixel.0).flatten().collect::<Vec<u8>>()
-            }).flatten().collect();
+                image
+                    .to_rgba8()
+                    .enumerate_pixels()
+                    .map(|(_, _, pixel)| pixel.0)
+                    .flatten()
+                    .collect::<Vec<u8>>()
+            })
+            .flatten()
+            .collect();
 
         eprintln!("writing...");
         encoder.write(&data)?;
